+++ RxCocoa/iOS/WKWebView+Rx.swift
@@ -1,5 +1,5 @@
 //
-//  UIWebView+Rx.swift
+//  WKWebView+Rx.swift
 //  RxCocoa
 //
 //  Created by Andrew Breckenridge on 8/30/16.
@@ -9,34 +9,35 @@
 #if os(iOS)
 
     import UIKit
+    import WebKit
     import RxSwift
 
-    extension Reactive where Base: UIWebView {
+    extension Reactive where Base: WKWebView {
 
         /// Reactive wrapper for `delegate`.
         /// For more information take a look at `DelegateProxyType` protocol documentation.
-        public var delegate: DelegateProxy<UIWebView, UIWebViewDelegate> {
+        public var delegate: DelegateProxy<WKWebView, WKNavigationDelegate> {
             return RxWebViewDelegateProxy.proxy(for: base)
         }
 
         /// Reactive wrapper for `delegate` message.
         public var didStartLoad: Observable<Void> {
             return delegate
-                .methodInvoked(#selector(UIWebViewDelegate.webViewDidStartLoad(_:)))
+                .methodInvoked(#selector(WKNavigationDelegate.webView(_:didStartProvisionalNavigation:)))
                 .map { _ in }
         }
 
         /// Reactive wrapper for `delegate` message.
         public var didFinishLoad: Observable<Void> {
             return delegate
-                .methodInvoked(#selector(UIWebViewDelegate.webViewDidFinishLoad(_:)))
+                .methodInvoked(#selector(WKNavigationDelegate.webView(_:didFinish:)))
                 .map { _ in }
         }
         
         /// Reactive wrapper for `delegate` message.
         public var didFailLoad: Observable<Error> {
             return delegate
-                .methodInvoked(#selector(UIWebViewDelegate.webView(_:didFailLoadWithError:)))
+                .methodInvoked(#selector(WKNavigationDelegate.webView(_:didFail:withError:)))
                 .map { a in
                     return try castOrThrow(Error.self, a[1])
                 }