+++ Tests/RxCocoaTests/WKWebView+RxTests.swift
@@ -1,5 +1,5 @@
 //
-//  UIWebView+RxTests.swift
+//  WKWebView+RxTests.swift
 //  Tests
 //
 //  Created by Andrew Breckenridge on 8/30/16.
@@ -9,54 +9,56 @@
 #if os(iOS)
     
 import UIKit
+import WebKit
 import RxSwift
 import RxCocoa
 import RxBlocking
 import XCTest
 
-final class UIWebViewTests: RxTest {}
+final class WKWebViewTests: RxTest {}
 
 fileprivate let testHTMLString = "<html><head></head><body><h1>🔥</h1></body></html>"
     
-extension UIWebViewTests {
+extension WKWebViewTests {
         
     func testDidStartLoad() {
-        let webView = UIWebView()
+        let webView = WKWebView()
         var didStartLoad = false
 
         let subscription = webView.rx.didStartLoad.subscribe(onNext: { _ in
             didStartLoad = true
         })
 
-        webView.delegate!.webViewDidStartLoad!(webView)
+        webView.navigationDelegate!.webView?(webView, didStartProvisionalNavigation: nil)
 
         XCTAssertTrue(didStartLoad)
         subscription.dispose()
     }
     
     func testDidFinishLoad() {
-        let webView = UIWebView()
+        let webView = WKWebView()
         var didFinishLoad = false
 
         let subscription = webView.rx.didFinishLoad.subscribe(onNext: { _ in
             didFinishLoad = true
         })
 
-        webView.delegate!.webViewDidFinishLoad!(webView)
+        webView.navigationDelegate!.webView?(webView, didFinish: nil)
 
         XCTAssertTrue(didFinishLoad)
         subscription.dispose()
     }
 
     func testDidFailLoad() {
-        let webView = UIWebView()
+        let webView = WKWebView()
         var didFailLoad = false
 
         let subscription = webView.rx.didFailLoad.subscribe { _ in
             didFailLoad = true
         }
 
-        webView.delegate!.webView!(webView, didFailLoadWithError: NSError(domain: "", code: 0, userInfo: .none))
+        let error = NSError(domain: "", code: 0, userInfo: .none)
+        webView.navigationDelegate!.webView!(webView, didFail: nil, withError: error)
 
         XCTAssertTrue(didFailLoad)
         subscription.dispose()